--- a/lib/adapters/http.js
+++ b/lib/adapters/http.js
@@ -215,11 +215,13 @@ module.exports = function httpAdapter(config) {
         settle(resolve, reject, response);
       } else {
         var responseBuffer = [];
+        var contentLength = 0;
         stream.on('data', function handleStreamData(chunk) {
           responseBuffer.push(chunk);
+          contentLength += chunk.length;
 
           // make sure the content length is not over the maxContentLength if specified
-          if (config.maxContentLength > -1 && Buffer.concat(responseBuffer).length > config.maxContentLength) {
+          if (config.maxContentLength > -1 && contentLength > config.maxContentLength) {
             stream.destroy();
             reject(createError('maxContentLength size of ' + config.maxContentLength + ' exceeded',
               config, null, lastRequest));