Refines the logic to calculate the total size of the chunks received. The current logic uses a `Buffer.concat` call which could get expensive as the number of chunks gets larger.

The issue and the root cause is well described here #2829

Here are the benchmarks before and after the fix.
Before
```
withMaxContentLength => Success in 10 seconds.
withoutMaxContentLength => Success in 0 seconds.
```
After
```
withMaxContentLength => Success in 1 seconds.
withoutMaxContentLength => Success in 0 seconds.
```