@jasonsaayman 2021-09-05
Seems this has been solved in the lates master release. Please re-open should you feel this is a issue that must be addressed.